+++ sort_ans.c
@@ -7,7 +7,7 @@
 #define J(i) (((i>>8)-1) / ((i&255)+1))
 #define K(i) (((i>>8)-1) % ((i&255)+1))
 
-int sort(int* a, long i){
+int sort(int* a, int i){
   for(i = (i*i*256) + (i-1); i >= 256; i = i - 256){
     ORDER(a[J(i)], a[K(i)]);
   }
